Rename convertNodeInfoToLegacyFormat as P2P sanitize util
